+++ admin/src/main/scala/com/blinkbox/books/reading/admin/AppConfig.scala
@@ -1,12 +1,15 @@
-package com.blinkbox.books.reading
+package com.blinkbox.books.reading.admin
 
 import akka.util.Timeout
-import com.blinkbox.books.config.{DatabaseConfig, ApiConfig, RichConfig}
+import com.blinkbox.books.clients.ClientConfig
+import com.blinkbox.books.config.{ApiConfig, AuthClientConfig, DatabaseConfig, RichConfig}
 import com.typesafe.config.Config
 
 case class AppConfig(
   api: ApiConfig,
   db: DatabaseConfig,
+  auth: AuthClientConfig,
+  catalogue: ClientConfig,
   startTimeout: Timeout
 )
 
@@ -14,6 +17,9 @@ object AppConfig {
   def apply(config: Config): AppConfig = AppConfig(
     ApiConfig(config, "service.reading.api.admin"),
     DatabaseConfig(config, "service.reading.db"),
+    AuthClientConfig(config),
+    ClientConfig(config),
     Timeout(config.getFiniteDuration("service.reading.startTimeout"))
   )
 }
+