+++ common/src/main/scala/com/blinkbox/books/clients/catalogue/CatalogueService.scala
@@ -66,8 +66,8 @@ class DefaultCatalogueV1Service(client: Client)(implicit ec: ExecutionContext) e
   def getBulkBookInfo(isbns: List[String], userId: Int): Future[BulkBookInfo] = {
     if (isbns.isEmpty) { Future.successful(BulkBookInfo(0, List.empty[BookInfo])) }
     else {
-      val isbnQueryString = isbns.mkString(start = "id=", sep = "&id=", end = "")
-      val req = Get(client.config.url.withPath(Uri.Path("/catalogue/books")).withQuery(isbns.map("id" -> _): _*))
+      val requestUri = Uri(s"${client.config.url}/catalogue/books").withQuery(isbns.map("id" -> _): _*)
+      val req = Get(requestUri)
       client.dataRequest[BulkBookInfo](req, credentials = None).transform(identity, {
         case e: NotFoundException =>
           new CatalogueInfoMissingException(s"Catalogue does not have a book with the following isbns: $isbns", e)
@@ -91,12 +91,16 @@ class DefaultCatalogueV1Service(client: Client)(implicit ec: ExecutionContext) e
   }
 
   override def getBulkContributorInfo(contributorIds: List[String]): Future[BulkContributorInfo] = {
-    val queryString = contributorIds.map(id => s"id=${id}").foldRight("")((a,b) => s"${a}&${b}")
-    val req = Get(s"${client.config.url}/catalogue/contributors?$queryString")
-    client.dataRequest[BulkContributorInfo](req, credentials = None).transform(identity, {
-      case e: NotFoundException =>
-        new CatalogueInfoMissingException(s"Catalogue does not have a contributor with id: $contributorIds", e)
-    })
+    if (contributorIds.isEmpty) { Future.successful(BulkContributorInfo(0, List.empty[ContributorInfo])) }
+    else {
+      val requestUrl = Uri(s"${client.config.url}/catalogue/contributors").withQuery(contributorIds.map("id" -> _): _*)
+      val req = Get(requestUrl)
+
+      client.dataRequest[BulkContributorInfo](req, credentials = None).transform(identity, {
+        case e: NotFoundException =>
+          new CatalogueInfoMissingException(s"Catalogue does not have a contributor with id: $contributorIds", e)
+      })
+    }
   }
 
   private def buildBulkCatalogueInfo(bulkBookInfo: BulkBookInfo): Future[List[CatalogueInfo]] = {