+++ common/src/main/scala/com/blinkbox/books/reading/Models.scala
@@ -7,15 +7,23 @@ object NotStarted extends ReadingStatus
 object Reading extends ReadingStatus
 object Finished extends ReadingStatus
 
-case class ReadingPosition(cfi: Cfi, percentage: Int) {
+case class ReadingPosition(cfi: Option[Cfi], percentage: Int) {
   require(percentage >= 0 && percentage <= 100)
 }
 
 case class Cfi(value: String)
 
-sealed trait BookType
-object Full extends BookType
-object Sample extends BookType
+sealed trait Ownership extends Ordered[Ownership] {
+  val order: Int
+}
+object Owned extends Ownership {
+  override val order: Int = 1
+  override def compare(that: Ownership): Int = this.order - that.order
+}
+object Sample extends Ownership {
+  override val order: Int = 0
+  override def compare(that: Ownership): Int = this.order - that.order
+}
 
 sealed trait LinkType extends Relation
 case object CoverImage extends LinkType
@@ -24,3 +32,5 @@ case object SampleEpub extends LinkType
 sealed trait LibraryMediaLinkType extends LinkType
 case object EpubKey extends LibraryMediaLinkType
 case object FullEpub extends LibraryMediaLinkType
+
+class LibraryItemConflict(msg: String, cause: Throwable = null) extends Exception(msg, cause)