+++ CHANGELOG.md
@@ -1,5 +1,27 @@
 # Change log
 
+## 0.6.0 ([#23](https://git.mobcastdev.com/Agora/reading-service/pull/23) 2014-12-02 17:36:20)
+
+Admin add book
+
+### New feature
+
+- `Add full book` admin endpoint.
+
+## 0.5.1 ([#22](https://git.mobcastdev.com/Agora/reading-service/pull/22) 2014-11-28 18:47:09)
+
+Fix catalogue url
+
+A patch to fix the url that calls the catalogue service.
+
+## 0.5.0 ([#21](https://git.mobcastdev.com/Agora/reading-service/pull/21) 2014-11-28 10:45:18)
+
+Add full book endpoint spec
+
+### New feature
+
+- Add full book endpoint spec.
+
 ## 0.4.0 ([#20](https://git.mobcastdev.com/Agora/reading-service/pull/20) 2014-11-26 14:53:22)
 
 Added health endpoint